ENH: Refactor `FunctionalRegression`
