+++ pgmpy/parameter/__init__.py
@@ -1,5 +1,12 @@
+from pgmpy.parameter.bayesian._base import BasePyroRegression
+from pgmpy.parameter.bayesian.BayesianMCMCRegression import BayesianMCMCRegression
+from pgmpy.parameter.bayesian.BayesianSVIRegression import BayesianSVIRegression
+
 from ._base import BaseParameter
 
 __all__ = [
     "BaseParameter",
+    "BasePyroRegression",
+    "BayesianMCMCRegression",
+    "BayesianSVIRegression",
 ]
